+++ Application/SettingsController.h
@@ -1,5 +1,5 @@
 /*
-    Copyright (c) 2003-2024, Sveinbjorn Thordarson <sveinbjorn@sveinbjorn.org>
+    Copyright (c) 2003-2025, Sveinbjorn Thordarson <sveinbjorn@sveinbjorn.org>
     All rights reserved.
 
     Redistribution and use in source and binary forms, with or without modification,
@@ -28,9 +28,9 @@
     POSSIBILITY OF SUCH DAMAGE.
 */
 
-#import <Cocoa/Cocoa.h>
+@import Cocoa;
 
-@interface PrefsController : NSWindowController <NSMenuDelegate>
+@interface SettingsController : NSWindowController <NSMenuDelegate>
 
 + (NSDictionary *)defaultsDictionary;
 + (void)putCommandLineToolInstallStatusInTextField:(NSTextField *)textField;