+++ Shared/NSWorkspace+Additions.m
@@ -1,5 +1,5 @@
 /*
-    Copyright (c) 2003-2024, Sveinbjorn Thordarson <sveinbjorn@sveinbjorn.org>
+    Copyright (c) 2003-2025, Sveinbjorn Thordarson <sveinbjorn@sveinbjorn.org>
     All rights reserved.
 
     Redistribution and use in source and binary forms, with or without modification,
@@ -180,6 +180,17 @@
         return [NSString stringWithFormat:@"%.1f MB", size / 1048576.0];
     }    
     return [NSString stringWithFormat:@"%.1f GB", size / 1073741824.0];
+}
+
+- (BOOL)isFileEmptyAtPath:(NSString *)path {
+    NSError *error = nil;
+    // Initialize data object without loading it into memory
+    NSData *d = [NSData dataWithContentsOfFile:path
+                                       options:NSDataReadingMappedAlways
+                                         error:&error];
+    // If error we don't know and we return NO
+    // If we do and it's empty, well...
+    return (error != nil || [d length] == 0);
 }
 
 #pragma mark - Open With menu