+++ CHANGES.md
@@ -28,11 +28,12 @@
 
 ## Version history
 
-### For 5.4.2 - 24/04/2024
+### For 5.5.0 - 24/05/2025
 
 * Fixed bug where the argument settings window would lock up
 * Better support for Dark Mode
 * Platypus now requires macOS 10.13 or later
+* Preferences now named Settings according to new macOS convention
 
 ### For 5.4.1 - 22/10/2022